Hide internal AVPlayer subclass
To hide the real class without the need for another property, the player is externally exposed in a category, whose method signature returns a different type from the underlying inner property.

This also was a good excuse to better organize the controller public API in categories (at the expense of having to create the corresponding properties internally).