Improve livestream type detection on tvOS
On tvOS, DVR streams might sometimes incorrectly behave as livestreams, including with the bare AVPlayerViewController and a simple AVPlayer. In such cases, those were incorrectly reported as on-demand streams, which was leading to potential issues (most notably with the Akamai workaround ending playback if the playhead position reaches the end of the stream).

The problem does not seem to affect the iOS player.